+++ packages/varlet-ui/src/space/Space.tsx
@@ -61,7 +61,6 @@ export default defineComponent({
 
       const lastIndex = children.length - 1
       const spacers = children.map((child, index) => {
-        const width = direction === 'row' ? undefined : '100%'
         let margin = '0'
 
         if (direction === 'row') {
@@ -88,12 +87,12 @@ export default defineComponent({
           margin = `0 0 ${y}px 0`
         }
 
-        return <div style={{ margin, width }}>{child}</div>
+        return <div style={{ margin }}>{child}</div>
       })
 
       return (
         <div
-          class={classes(n(), 'var--box', [inline, n('--inline')])}
+          class={classes(n(), n('$--box'), [inline, n('--inline')])}
           style={{
             flexDirection: direction,
             justifyContent: padStartFlex(justify),